Fix overflow on zariths, add test

+++ src/Tezos/Forging.cpp
@@ -8,6 +8,7 @@
 #include "../Base58.h"
 #include "../Data.h"
 #include "../HexCoding.h"
+#include "../BinaryCoding.h"
 
 #include <sstream>
 
@@ -60,7 +61,7 @@ Data forgePublicKey(PublicKey publicKey) {
 }
 
 // Forge the given zarith hash into a hex encoded string.
-Data forgeZarith(uint16_t input) {
+Data forgeZarith(int input) {
     Data forged = Data();
     while (input > 0x80) {
         forged.push_back(static_cast<byte>((input & 0xff) | 0x80));
